propagate exceptions from search functions, and fix error detection
Corrected the error detection in pyipopt.solve so that it properly
throws exceptions when there are problems. Additionally, record and
propagate exceptions from the inner search functions such as the
objective and gradient functions. This also required a large
refactoring of the callback code to stop duplicating the cleanup code
in the callback stubs.

* hook.h: Added save_python_exception declaration
* callback.c: Cleaned up callback stubs to unify the error detection
  and cleanup, as well as save and python exceptions during cleanup
* pyipopt.c: Fixed error detection in solve, and restore python
  exceptions on errors